requirements/base.txt
pyzbar==0.1.8
Pillow==6.1.0
python-telegram-bot==11.1.0